refactor(cli): add runner trait
