+++ crates/oxc_cli/src/lint/isolated_handler.rs
@@ -16,17 +16,16 @@ use oxc_diagnostics::{
     thiserror::Error,
     Error, GraphicalReportHandler, Severity,
 };
-use oxc_linter::{Fixer, LintContext, Linter, RuleCategory, RuleEnum, RULES};
+use oxc_linter::{Fixer, LintContext, Linter};
 use oxc_parser::Parser;
 use oxc_semantic::SemanticBuilder;
 use oxc_span::SourceType;
-use rustc_hash::FxHashSet;
 
-use super::{AllowWarnDeny, LintOptions};
+use super::options::LintOptions;
 use crate::{CliRunResult, Walk};
 
-pub struct LintRunner {
-    options: LintOptions,
+pub struct IsolatedLintHandler {
+    options: Arc<LintOptions>,
 
     linter: Arc<Linter>,
 }
@@ -36,67 +35,15 @@ pub struct LintRunner {
 #[diagnostic(help("{0:?} seems like a minified file"))]
 pub struct MinifiedFileError(pub PathBuf);
 
-impl LintRunner {
-    pub fn new(options: LintOptions) -> Self {
-        let linter = Linter::from_rules(Self::derive_rules(&options)).with_fix(options.fix);
-        Self { options, linter: Arc::new(linter) }
-    }
-
-    pub fn print_rules() {
-        let mut stdout = BufWriter::new(std::io::stdout());
-        Linter::print_rules(&mut stdout);
-    }
-
-    fn derive_rules(options: &LintOptions) -> Vec<RuleEnum> {
-        let mut rules: FxHashSet<RuleEnum> = FxHashSet::default();
-
-        for (allow_warn_deny, name_or_category) in &options.rules {
-            let maybe_category = RuleCategory::from(name_or_category.as_str());
-            match allow_warn_deny {
-                AllowWarnDeny::Deny => {
-                    match maybe_category {
-                        Some(category) => rules.extend(
-                            RULES.iter().filter(|rule| rule.category() == category).cloned(),
-                        ),
-                        None => {
-                            if name_or_category == "all" {
-                                rules.extend(RULES.iter().cloned());
-                            } else {
-                                rules.extend(
-                                    RULES
-                                        .iter()
-                                        .filter(|rule| rule.name() == name_or_category)
-                                        .cloned(),
-                                );
-                            }
-                        }
-                    };
-                }
-                AllowWarnDeny::Allow => {
-                    match maybe_category {
-                        Some(category) => rules.retain(|rule| rule.category() != category),
-                        None => {
-                            if name_or_category == "all" {
-                                rules.clear();
-                            } else {
-                                rules.retain(|rule| rule.name() == name_or_category);
-                            }
-                        }
-                    };
-                }
-            }
-        }
-
-        let mut rules = rules.into_iter().collect::<Vec<_>>();
-        // for stable diagnostics output ordering
-        rules.sort_unstable_by_key(|rule| rule.name());
-        rules
+impl IsolatedLintHandler {
+    pub(super) fn new(options: Arc<LintOptions>, linter: Arc<Linter>) -> Self {
+        Self { options, linter }
     }
 
     /// # Panics
     ///
     /// * When `mpsc::channel` fails to send.
-    pub fn run(&self) -> CliRunResult {
+    pub(super) fn run(&self) -> CliRunResult {
         let now = std::time::Instant::now();
 
         let number_of_files = Arc::new(AtomicUsize::new(0));
